requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.41

# AI Providers
anthropic==0.18.1
openai==1.10.0

# Slack Integration
slack-sdk==3.26.2

# Terraform Validation
python-hcl2==4.3.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0
python-json-logger==2.0.7

# Security
cryptography==42.0.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Kubernetes
kubernetes==29.0.0

# Monitoring
prometheus-client==0.19.0
